# Use UglifyJS when possible, but fall back to Babel.minify.