--- a/packages/minifier-js/.npm/package/.gitignore
+++ b/packages/minifier-js/.npm/package/.gitignore
@@ -1,0 +1,1 @@
+node_modules