--- a/packages/minifier-js/package.js
+++ b/packages/minifier-js/package.js
@@ -3,8 +3,12 @@ Package.describe({
   version: "2.0.0-rc.1"
 });
 
+Npm.depends({
+  "uglify-js": "2.8.16"
+});
+
 Package.onUse(function (api) {
   api.use('babel-compiler');
-  api.export(['meteorBabelMinify']);
+  api.export(['meteorJsMinify']);
   api.addFiles(['minifier.js'], 'server');
 });