… packages/minifier-js/.npm/package/npm-shrinkwrap.json